Migrate from gometalinter to golangci
fixes https://github.com/hazelcast/hazelcast-go-client/issues/458

+++ test/proxy/pncounter/pn_counter_test.go
@@ -258,5 +258,5 @@ func TestPNCounter_Reset(t *testing.T) {
 	counter.Reset()
 	currentValue, err := counter.AddAndGet(delta)
 	require.NoError(t, err)
-	assert.Equalf(t, currentValue, int64(delta), "PNCounter.Reset failed")
+	assert.Equalf(t, currentValue, delta, "PNCounter.Reset failed")
 }
